Refactor utils into sub-modules
