+++ src/gains/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility functionality, used across various package (sub)modules."""
