+++ src/gains/utils/profile.py
@@ -1,0 +1,7 @@
+"""Wrappers and decorators for profiling runs."""
+
+import argparse
+
+
+def add_profiling_options(parser: argparse.ArgumentParser) -> None:
+    """Adds necessary profiling options to ArgumentParser instances."""
